Most Panini reachable forms of लेट् supported
