+++ vidyut-prakriya/tests/integration/main.rs
@@ -66,4 +66,7 @@ mod dhaturatnakara;
 
 mod macdonell;
 
+mod rv_let_aorist_corpus;
 mod rv_let_corpus;
+mod rv_let_helpers;
+mod rv_let_perfect_corpus;
